move functions to outside form component
