+++ src/applications/coronavirus-screener/components/MultiQuestionForm.jsx
@@ -1,66 +1,17 @@
 /* eslint-disable no-console */
 import React, { useState, useEffect } from 'react';
-import { Element, scroller } from 'react-scroll';
+import { Element } from 'react-scroll';
 import FormQuestion from './FormQuestion';
 import FormResult from './FormResult';
 import recordEvent from 'platform/monitoring/record-event';
 import moment from 'moment';
 import _ from 'lodash/fp';
-
-// scoller usage based on https://github.com/department-of-veterans-affairs/veteran-facing-services-tools/blob/master/packages/formation-react/src/components/CollapsiblePanel/CollapsiblePanel.jsx
-
-function scrollTo(name) {
-  scroller.scrollTo(
-    name,
-    window.VetsGov.scroll || {
-      duration: 500,
-      delay: 2,
-      smooth: true,
-    },
-  );
-}
-
-function getEnabledQuestions(questionState) {
-  return questionState.filter(question => question.enabled ?? true);
-}
-
-// check if all enabled questions have been answered
-function checkFormComplete(questionState) {
-  const completedQuestions = getEnabledQuestions(questionState).map(question =>
-    Object.prototype.hasOwnProperty.call(question, 'value'),
-  );
-  return !completedQuestions.includes(false);
-}
-
-// check result of answers
-function checkFormResult(questionState) {
-  return getEnabledQuestions(questionState)
-    .map(question => {
-      const passValues = question.passValues ?? ['no'];
-      return passValues.includes(question.value);
-    })
-    .includes(false)
-    ? 'fail'
-    : 'pass';
-}
-
-// check the overall status of the form
-function checkFormStatus(questionState) {
-  return !checkFormComplete(questionState)
-    ? 'incomplete'
-    : checkFormResult(questionState);
-}
-
-function recordCompletion({ formState, setFormState }) {
-  if (formState.status !== 'incomplete' && formState.completed === false) {
-    recordEvent({
-      event: 'covid-screening-tool-result-displayed',
-      'screening-tool-result': formState.result,
-      'time-to-complete': moment().unix() - formState.startTime,
-    });
-    setFormState({ ...formState, completed: true });
-  }
-}
+import {
+  getEnabledQuestions,
+  checkFormStatus,
+  recordCompletion,
+  scrollTo,
+} from '../lib';
 
 export default function MultiQuestionForm({ questions, defaultOptions }) {
   const [formState, setFormState] = useState({
@@ -115,7 +66,7 @@ export default function MultiQuestionForm({ questions, defaultOptions }) {
   );
 
   // records startTime and log to GA
-  const recordStart = question => {
+  function recordStart(question) {
     if (formState.startTime === null) {
       recordEvent({
         event: 'covid-screening-tool-start',
@@ -127,7 +78,7 @@ export default function MultiQuestionForm({ questions, defaultOptions }) {
         startTime: moment().unix(),
       });
     }
-  };
+  }
 
   const enabledQuestions = getEnabledQuestions(questionState);
